Refactor the way the project comparator is called after the manipulation
Introduces a new interface, Callback, that can be used to extend the
behavior of the code. Rework the project comparator so that it implements
such interface in order to decouple the post alignment behavior from the
comparator code, thus allowing us to implement additional post-alignment
behaviors.
