+++ cli/src/main/resources/logback.xml
@@ -35,7 +35,7 @@
       <pattern>%mdc{LOG-CONTEXT}%logger{0} %msg%n</pattern>
     </encoder>
   </appender>
-  <logger name="org.commonjava.maven.ext.common.util.ProjectComparator" level="INFO" additivity="false">
+  <logger name="org.commonjava.maven.ext.common.callbacks.ComparatorCallback" level="INFO" additivity="false">
     <appender-ref ref="COMPARATOR" />
   </logger>
